// src/app/models/user.module.ts
export class Users {
  //personalDetail:{
  _id: number;
  fullname: string;
  gender: string;
  motherTongue: string;
  Clause: string;
  dob: Date;
  birthTime: string;
  age: number;
  height: number;
  color: string;
  qualificaton: string;
  occupation: string;
  work: string;
  earnings: string;
  physicalFitness: string;
  maritalStatus: string;
  religion: string;
  familyGod: string;
  //  }
  //familyDetail:{
  fathername: string;
  mothername: string;
  fatherOccupation: string;
  motherOccupation: string;
  //sibblings:{
  sibblingsMale: number;
  sibblingsFemale: number;
  //  },
  //marriedSibblings:{
  marriedSibblingsMale: number;
  marriedSibblingsFemale: number;
  //},
  origin: string;
  native: string;
  assets: string;
  dowry: string;
  //}
  //horoscopeDetails:{
  star: string;
  rasi: string;
  lagnam: string;
  directionBias: string;
  year: string;
  month: string;
  day: string;
  //rasibox:{
  ra1: string;
  ra2: string;
  ra3: string;
  ra4: string;
  ra5: string;
  ra6: string;
  ra7: string;
  ra8: string;
  ra9: string;
  ra10: string;
  ra11: string;
  ra12: string;
  //},
  //feature:{
  f1: string;
  f2: string;
  f3: string;
  f4: string;
  f5: string;
  f6: string;
  f7: string;
  f8: string;
  f9: string;
  f10: string;
  f11: string;
  f12: string;
  //}
  address: string;
  district: string;
  phone1: number;
  phone2: number;
  mailid: string;
  password: string;
  expectations: string;
  otherMessage: string;
  userImage: File;
  //}
}